日本語名追加

+++ EFT_Assistant/Project/source/Parts/StrControl.hpp
@@ -62,13 +62,13 @@ namespace FPS_n2 {
 	//
 	static const auto		UTF8toSjis(std::string srcUTF8) {
 		//Unicodeへ変換後の文字列長を得る
-		int lenghtUnicode = MultiByteToWideChar(CP_UTF8, 0, srcUTF8.c_str(), srcUTF8.size() + 1, NULL, 0);
+		int lenghtUnicode = (int)MultiByteToWideChar(CP_UTF8, 0, srcUTF8.c_str(), (int)(srcUTF8.size()) + 1, NULL, 0);
 
 		//必要な分だけUnicode文字列のバッファを確保
 		wchar_t* bufUnicode = new wchar_t[lenghtUnicode];
 
 		//UTF8からUnicodeへ変換
-		MultiByteToWideChar(CP_UTF8, 0, srcUTF8.c_str(), srcUTF8.size() + 1, bufUnicode, lenghtUnicode);
+		MultiByteToWideChar(CP_UTF8, 0, srcUTF8.c_str(), (int)(srcUTF8.size()) + 1, bufUnicode, lenghtUnicode);
 
 		//ShiftJISへ変換後の文字列長を得る
 		int lengthSJis = WideCharToMultiByte(CP_THREAD_ACP, 0, bufUnicode, -1, NULL, 0, NULL, NULL);
